Extract PythonSamplerAdapter into its own module
