+++ rustuna_pyo3/src/study.rs
@@ -21,8 +21,9 @@ use crate::attrs::{convert_pydict_to_fixed_params, pyobj_to_attrs_with_kind, Att
 use crate::exception::err_to_exceptions;
 use crate::pyobject_storage::PyObjectStorage;
 use crate::sampler::cmaes::PyCmaEsSampler;
+use crate::sampler::python::PythonSamplerAdapter;
 use crate::sampler::random::PyRandomSampler;
-use crate::sampler::{PyObjectSampler, PySampler};
+use crate::sampler::PySampler;
 use crate::storage::PyStorage;
 use crate::trial::{PyPersistedTrial, PyTrial, PyTrialState};
 use crate::trial_queue::{PyObjectTrialQueue, PyPyObjectTrialQueue, PyTrialQueue};
@@ -187,7 +188,7 @@ fn resolve_sampler_pyobj(
     } else if let Ok(py_random_sampler) = sampler_ref.extract::<PyRandomSampler>() {
         Ok((py_random_sampler.sampler.clone(), sampler_pyobj))
     } else {
-        let sampler: SharedSampler = Arc::new(Mutex::new(PyObjectSampler::new(sampler)));
+        let sampler: SharedSampler = Arc::new(Mutex::new(PythonSamplerAdapter::new(sampler)));
         Ok((sampler, sampler_pyobj))
     }
 }
